push updates

+++ utils/utils.py
@@ -1,5 +1,6 @@
 from openpyxl import load_workbook
 from openpyxl.utils import get_column_letter
+from openpyxl.worksheet.table import Table, TableStyleInfo
 from datetime import datetime
 import requests
 import os
@@ -10,7 +11,7 @@ from sqlalchemy.orm import sessionmaker
 
 def _findRowIndex(ws, searchTerm, column_id=1, limit=1e3):
     index = None
-    
+
     row_id = 1
     while row_id < limit:
         if ws.cell(row_id,column_id).value == searchTerm:
@@ -23,7 +24,7 @@ def _findRowIndex(ws, searchTerm, column_id=1, limit=1e3):
 
 def _findColumnIndex(ws, searchTerm, row_id=1, limit=1e3):
     index = None
-    
+
     column_id = 1
     while column_id < limit:
         if ws.cell(row_id,column_id).value == searchTerm:
@@ -43,34 +44,34 @@ def _removeRows(ws, searchTerm, row_id=1, column_id=1, limit=1e3):
     return
 
 
-def _getIssue22OfCommune(wb, commune_id, issue22_list):
-    ws = wb.create_sheet('ISSUE22')
-    currentLine_idx = 1
+# def _getIssue22OfCommune(wb, commune_id, issue22_list):
+#     ws = wb.create_sheet('ISSUE22')
+#     currentLine_idx = 1
 
-    if issue22_list == []:
-        return
+#     if issue22_list == []:
+#         return
 
-    # title line
-    for j, title in enumerate(issue22_list[0].keys()):
-        ws.cell(currentLine_idx,j+1).value = title
+#     # title line
+#     for j, title in enumerate(issue22_list[0].keys()):
+#         ws.cell(currentLine_idx,j+1).value = title
 
-    currentLine_idx += 1
+#     currentLine_idx += 1
 
-    for i22 in issue22_list:
-        if i22['COM_FOSNR'] == commune_id:
-            ws.cell(currentLine_idx,1).value = i22['COM_FOSNR']
-            ws.cell(currentLine_idx,2).value = i22['AV_SOURCE']
-            ws.cell(currentLine_idx,3).value = i22['AV_TYPE']
-            ws.cell(currentLine_idx,4).value = i22['ISSUE']
-            ws.cell(currentLine_idx,5).value = i22['ISSUE_CATEGORY']
-            ws.cell(currentLine_idx,6).value = i22['BDG_E']
-            ws.cell(currentLine_idx,7).value = i22['BDG_N']
-            ws.cell(currentLine_idx,8).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(i22['BDG_E'],i22['BDG_N'])
-            ws.cell(currentLine_idx,8).value = 'sitn.ne.ch'
-            ws.cell(currentLine_idx,8).style = 'Hyperlink'
-            currentLine_idx += 1
+#     for i22 in issue22_list:
+#         if i22['COM_FOSNR'] == commune_id:
+#             ws.cell(currentLine_idx,1).value = i22['COM_FOSNR']
+#             ws.cell(currentLine_idx,2).value = i22['AV_SOURCE']
+#             ws.cell(currentLine_idx,3).value = i22['AV_TYPE']
+#             ws.cell(currentLine_idx,4).value = i22['ISSUE']
+#             ws.cell(currentLine_idx,5).value = i22['ISSUE_CATEGORY']
+#             ws.cell(currentLine_idx,6).value = i22['BDG_E']
+#             ws.cell(currentLine_idx,7).value = i22['BDG_N']
+#             ws.cell(currentLine_idx,8).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(i22['BDG_E'],i22['BDG_N'])
+#             ws.cell(currentLine_idx,8).value = 'sitn.ne.ch'
+#             ws.cell(currentLine_idx,8).style = 'Hyperlink'
+#             currentLine_idx += 1
 
-    return wb
+#     return wb
 
 
 def _get_issue(error, issue_solution):
@@ -115,7 +116,7 @@ def loadCommunesOFS():
     i = 0
     while i < 1e4:
         i += 1
-        
+
         if ws_source.cell(i,1).value is None:
             break
 
@@ -127,24 +128,25 @@ def loadCommunesOFS():
     return communes_ofs
 
 
-def downloadListeCantonNeuchatel(path=None):
+def downloadListeCantonNeuchatel(path=None, batprojtreat=False):
     if path is None or not os.path.exists(path):
         path = os.environ['FEEDBACK_COMMUNES_WORKING_DIR']
 
     r = requests.get(os.environ['STATIC_URL_FEEBDACK_CANTON'], allow_redirects=True)
 
-    filename = datetime.strftime(datetime.now(), '%Y%M%d_Listes_NE.xlsx')
+    filename = datetime.strftime(datetime.now(), '%Y%m%d_Listes_NE.xlsx')
 
     filepath = os.path.join(path, filename)
     open(filepath, 'wb').write(r.content)
-    tempfilepath = os.path.join(path, 'temp_' + filename)
-    open(tempfilepath, 'wb').write(r.content)
+    if batprojtreat is True:
+        tempfilepath = os.path.join(path, 'temp_' + filename)
+        open(tempfilepath, 'wb').write(r.content)
 
-    fme_command = 'fme "{}" --SourceDataset_POSTGIS "{}" --SourceDataset_XLSXR "{}" --DestDataset_XLSXW "{}"'.format(os.environ['FEEDBACK_COMMUNES_LISTES_BATPROJ_FILTER_FME_PATH'], os.environ['FEEDBACK_COMMUNES_LISTES_BATPROJ_FILTER_FME_USER'], tempfilepath, filepath)
-    print('\ncommand: {}"\n'.format(fme_command))
-    os.system(fme_command)
+        fme_command = 'fme "{}" --SourceDataset_POSTGIS "{}" --SourceDataset_XLSXR "{}" --DestDataset_XLSXW "{}"'.format(os.environ['FEEDBACK_COMMUNES_LISTES_BATPROJ_FILTER_FME_PATH'], os.environ['FEEDBACK_COMMUNES_LISTES_BATPROJ_FILTER_FME_USER'], tempfilepath, filepath)
+        print('\ncommand: {}"\n'.format(fme_command))
+        os.system(fme_command)
 
-    os.remove(tempfilepath)
+        os.remove(tempfilepath)
 
     return filepath
 
@@ -155,7 +157,7 @@ def downloadIssue22CantonNeuchatel(path=None):
 
     r = requests.get(os.environ['FEEDBACK_COMMUNES_URL_ISSUE_22_CANTON'], allow_redirects=True)
 
-    filename = datetime.strftime(datetime.now(), '%Y%M%d_issue22_NE.xlsx')
+    filename = datetime.strftime(datetime.now(), '%Y%m%d_issue22_NE.xlsx')
 
     filepath = os.path.join(path, filename)
     open(filepath, 'wb').write(r.content)
@@ -171,7 +173,7 @@ def downloadIssue22CantonNeuchatel(path=None):
 
     wb = load_workbook(filepath)
     ws = wb['NE']
-    
+
     issue22_list = []
     i = 1
     while i < 1e5 and ws.cell(i,1) is not None:
@@ -198,19 +200,32 @@ def cleanWorkingDirectory(path=None):
     return
 
 
+def _createExcelTable(ws, table_name, ref):
+    tab = Table(displayName=table_name, ref=ref)
+
+    style = TableStyleInfo(name="TableStyleMedium9", showFirstColumn=False,
+                        showLastColumn=False, showRowStripes=True, showColumnStripes=False)
+
+    tab.tableStyleInfo = style
+
+    ws.add_table(tab)
+
+    return
+
+
 def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath, issue22_list, issue_solution, today=datetime.strftime(datetime.now(), '%Y%m%d'), environ='INTER'):
     # copy canton_file to commune_file
     feedback_commune_filename = '_'.join([str(commune_id), commune_name.replace(' ', '_'), 'feedback', today]) + '.xlsx'
     feedback_commune_filepath = os.path.join(os.environ['FEEDBACK_COMMUNES_WORKING_DIR'], today, feedback_commune_filename)
     shutil.copy2(feedback_canton_filepath, feedback_commune_filepath)
-    
+
     feedback_commune = {
         'commune_id': commune_id,
         'commune_nom': commune_name
     }
 
     wb = load_workbook(feedback_commune_filepath)
-    
+
     # remove canton sheet
     wb.remove(wb['Cantons'])
 
@@ -219,7 +234,7 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
     row_i = _findRowIndex(ws, 'Commune', column_id=4)
     _removeRows(ws, commune_id, row_id=row_i+2, column_id=3, limit=1e2)
 
- 
+
     # Create resume sheet
     ws2 = wb.create_sheet('resume')
 
@@ -289,11 +304,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
+
         ws2.cell(ws2_line_i,1).value = 'LISTE 1 - Bâtiments sans coordonnées'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'STRNAME'
@@ -312,8 +328,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2.cell(ws2_line_i,5).value = ws.cell(ws_line_i,15).value
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
+
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(5) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste1', ref)
 
     #####################
     #  LISTE 2
@@ -323,11 +343,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
+
         ws2.cell(ws2_line_i,1).value = 'LISTE 2 - Coordonnées en dehors de la commune'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'Adresse'
@@ -337,17 +358,21 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
         ws2_line_i += 1
 
         while ws.cell(ws_line_i,2).value is not None:
-            if ws.cell(ws_line_i,2).value == commune_id and ws.cell(ws_line_i, 18).value != 'bat_proj':
+            if ws.cell(ws_line_i,2).value == commune_id:
                 ws2.cell(ws2_line_i,1).value = ws.cell(ws_line_i,4).value
-                ws2.cell(ws2_line_i,1).hyperlink = os.environ['RAPPORT_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,11).value, ws.cell(ws_line_i,12).value)
+                ws2.cell(ws2_line_i,1).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,11).value, ws.cell(ws_line_i,12).value)
                 ws2.cell(ws2_line_i,1).style = 'Hyperlink'
                 ws2.cell(ws2_line_i,2).value = ws.cell(ws_line_i,5).value
                 ws2.cell(ws2_line_i,3).value = ws.cell(ws_line_i,11).value
                 ws2.cell(ws2_line_i,4).value = ws.cell(ws_line_i,12).value
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
+
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(4) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste2', ref)
 
     #####################
     #  LISTE 3
@@ -357,11 +382,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
+
         ws2.cell(ws2_line_i,1).value = 'LISTE 3 - Divergence de NPA'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'PLZ4 RegBL'
@@ -372,9 +398,9 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
         ws2_line_i += 1
 
         while ws.cell(ws_line_i,2).value is not None:
-            if ws.cell(ws_line_i,2).value == commune_id and ws.cell(ws_line_i, 26).value != 'bat_proj':
+            if ws.cell(ws_line_i,2).value == commune_id:
                 ws2.cell(ws2_line_i,1).value = ws.cell(ws_line_i,4).value
-                ws2.cell(ws2_line_i,1).hyperlink = os.environ['RAPPORT_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,20).value, ws.cell(ws_line_i,21).value)
+                ws2.cell(ws2_line_i,1).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,20).value, ws.cell(ws_line_i,21).value)
                 ws2.cell(ws2_line_i,1).style = 'Hyperlink'
                 ws2.cell(ws2_line_i,2).value = ws.cell(ws_line_i,8).value
                 ws2.cell(ws2_line_i,3).value = ws.cell(ws_line_i,9).value
@@ -382,8 +408,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2.cell(ws2_line_i,5).value = ws.cell(ws_line_i,12).value
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
+
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(5) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste3', ref)
 
     #####################
     #  LISTE 4
@@ -393,11 +423,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
-        ws2.cell(ws2_line_i,1).value = 'LISTE 4 - Bâtiments sans usage d\'habitation (déjà dans le RegBL)'
+
+        ws2.cell(ws2_line_i,1).value = 'LISTE 4 - Doublets d\'adresses'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'GKAT'
@@ -412,9 +443,9 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
         ws2_line_i += 1
 
         while ws.cell(ws_line_i,2).value is not None:
-            if ws.cell(ws_line_i,2).value == commune_id and ws.cell(ws_line_i, 25).value != 'bat_proj':
+            if ws.cell(ws_line_i,2).value == commune_id:
                 ws2.cell(ws2_line_i,1).value = ws.cell(ws_line_i,4).value
-                ws2.cell(ws2_line_i,1).hyperlink = os.environ['RAPPORT_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,7).value, ws.cell(ws_line_i,8).value)
+                ws2.cell(ws2_line_i,1).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(ws.cell(ws_line_i,7).value, ws.cell(ws_line_i,8).value)
                 ws2.cell(ws2_line_i,1).style = 'Hyperlink'
                 ws2.cell(ws2_line_i,2).value = ws.cell(ws_line_i,6).value
                 ws2.cell(ws2_line_i,3).value = ws.cell(ws_line_i,22).value
@@ -426,9 +457,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2.cell(ws2_line_i,9).value = ws.cell(ws_line_i,24).value
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
 
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(9) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste4', ref)
 
     #####################
     #  LISTE 5
@@ -438,11 +472,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
+
         ws2.cell(ws2_line_i,1).value = 'LISTE 5 - Définition du bâtiment'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'GKAT'
@@ -459,7 +494,7 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
             if ws.cell(ws_line_i,2).value == commune_id and ws.cell(ws_line_i, 13).value != 'bat_proj':
                 ws2.cell(ws2_line_i,1).value = ws.cell(ws_line_i,4).value
                 (coord_e, coord_n) = ws.cell(ws_line_i,9).value.split(' ')
-                ws2.cell(ws2_line_i,1).hyperlink = os.environ['RAPPORT_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(coord_e, coord_n)
+                ws2.cell(ws2_line_i,1).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(coord_e, coord_n)
                 ws2.cell(ws2_line_i,1).style = 'Hyperlink'
                 ws2.cell(ws2_line_i,2).value = ws.cell(ws_line_i,5).value
                 ws2.cell(ws2_line_i,3).value = ws.cell(ws_line_i,6).value
@@ -470,8 +505,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2.cell(ws2_line_i,8).value = _get_issue(ws.cell(ws_line_i,12).value, issue_solution)
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
+
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(8) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste5', ref)
 
     #####################
     #  LISTE 6
@@ -481,11 +520,12 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
 
     if ws_line_i is not None:
         ws2_line_i += 1
-        
+
         ws2.cell(ws2_line_i,1).value = 'Liste 6 - Catégorie du bâtiment'
         ws2.cell(ws2_line_i,1).style = 'Headline 1'
 
         ws2_line_i += 1
+        table_start_row = ws2_line_i
 
         ws2.cell(ws2_line_i,1).value = 'EGID'
         ws2.cell(ws2_line_i,2).value = 'GKAT'
@@ -502,7 +542,7 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
             if ws.cell(ws_line_i,2).value == commune_id and ws.cell(ws_line_i, 13).value != 'bat_proj':
                 ws2.cell(ws2_line_i,1).value = ws.cell(ws_line_i,4).value
                 (coord_e, coord_n) = ws.cell(ws_line_i,9).value.split(' ')
-                ws2.cell(ws2_line_i,1).hyperlink = os.environ['RAPPORT_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(coord_e, coord_n)
+                ws2.cell(ws2_line_i,1).hyperlink = os.environ['FEEDBACK_COMMUNES_URL_CONSULTATION_' + environ + '_ISSUE_22_SITN_COORD'].format(coord_e, coord_n)
                 ws2.cell(ws2_line_i,1).style = 'Hyperlink'
                 ws2.cell(ws2_line_i,2).value = ws.cell(ws_line_i,5).value
                 ws2.cell(ws2_line_i,3).value = ws.cell(ws_line_i,6).value
@@ -513,9 +553,13 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2.cell(ws2_line_i,8).value = _get_issue(ws.cell(ws_line_i,12).value, issue_solution)
 
                 ws2_line_i += 1
-        
+
             ws_line_i += 1
-    
+
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(8) + str(table_end_row)
+        _createExcelTable(ws2, 'Liste6', ref)
+
     #####################
     #  ISSUE 22
     #####################
@@ -526,18 +570,19 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
                 ws2_line_i += 1
                 ws2.cell(ws2_line_i,1).value = 'Bâtiments manquants'
                 ws2.cell(ws2_line_i,1).style = 'Headline 1'
-                
+
                 ws2_line_i += 1
+                table_start_row = ws2_line_i
 
                 ws2.cell(ws2_line_i,1).value = 'COORDE'
                 ws2.cell(ws2_line_i,2).value = 'COORDN'
                 ws2.cell(ws2_line_i,3).value = 'LINK'
                 ws2.cell(ws2_line_i,4).value = 'DESIGNATION_MO'
-                
+
                 ws2_line_i += 1
-                
+
                 anyI22 = True
-            
+
             ws2.cell(ws2_line_i,1).value = i22['BDG_E']
             ws2.cell(ws2_line_i,2).value = i22['BDG_N']
             ws2.cell(ws2_line_i,3).value = 'sitn.ne.ch'
@@ -545,8 +590,13 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
             ws2.cell(ws2_line_i,3).style = 'Hyperlink'
             ws2.cell(ws2_line_i,4).value = i22['DESIGNATION_MO']
 
-            
+
             ws2_line_i += 1
+
+    if anyI22 is True:
+        table_end_row = ws2_line_i-1
+        ref = 'A' + str(table_start_row) + ":" + get_column_letter(4) + str(table_end_row)
+        _createExcelTable(ws2, 'issue22', ref)
 
 
     # ajuster la largeur des colonnes
@@ -556,7 +606,7 @@ def generateCommuneErrorFile(commune_id, commune_name, feedback_canton_filepath,
     # supprimer toutes les colonnes autres que "resume"
     for ws_name in wb.sheetnames:
         if not ws_name == 'resume':
-            wb.remove(wb[ws_name])            
+            wb.remove(wb[ws_name])
 
     wb.save(feedback_commune_filepath)
